Prepare for custom functions
