+++ snuba/query/parser/exceptions.py
@@ -1,21 +1,7 @@
-from snuba.query.expressions import Expression
-
-
-class InvalidQueryException(Exception):
-    """
-    Common parent class used for invalid queries during parsing
-    and validation.
-    This should not be used for system errors.
-    """
-
-    pass
+from snuba.query.exceptions import InvalidQueryException, ValidationException
 
 
 class ParsingException(InvalidQueryException):
-    pass
-
-
-class ValidationException(InvalidQueryException):
     pass
 
 
@@ -25,13 +11,3 @@ class CyclicAliasException(ValidationException):
 
 class AliasShadowingException(ValidationException):
     pass
-
-
-class InvalidExpressionException(ValidationException):
-    def __init__(self, expression: Expression, message: str) -> None:
-        self.expression = expression
-        self.message = message
-        super().__init__(message)
-
-    def __str__(self) -> str:
-        return f"Invalid Expression {self.expression}: {self.message}"
